requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
blobfile==2.1.1
cached-property==1.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
chex==0.1.6
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
DALL-E==0.1
dalle-mini==0.1.5
decorator==4.4.2
depth_anything @ https://github.com/huchenlei/Depth-Anything/releases/download/v1.0.0/depth_anything-2024.1.22.0-py2.py3-none-any.whl#sha256=26c1d38b8c3c306b4a2197d725a4b989ff65f7ebcf4fb5a96a1b6db7fbd56780
diffusers==0.27.2
dm-tree==0.1.8
docker-pycreds==0.4.0
einops==0.7.0
emoji==2.11.0
etils==1.7.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.13.4
flatbuffers==24.3.25
flax==0.6.3
fonttools==4.51.0
freetype-py==2.4.0
fsspec==2024.3.1
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
gradio==3.50.2
gradio_client==0.6.1
h11==0.14.0
hsluv==5.0.4
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipython==8.23.0
jax
jaxlib
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.42.0
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
min-dalle==0.4.11
ml-dtypes==0.4.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
mypy==1.9.0
mypy-extensions==1.0.0
networkx==3.3
numba==0.59.1
numba-progress==1.1.0
numpy==1.26.4
onnxruntime==1.17.3
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optax==0.1.7
orbax==0.0.23
orjson==3.10.1
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.4.0
pooch==1.8.1
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pycryptodomex==3.20.0
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
Pygments==2.17.2
PyMatting==1.1.12
pyparsing==3.1.2
PyQt6==6.6.1
PyQt6-Qt6==6.6.3
PyQt6-sip==13.6.0
pyreadline3==3.4.1
pytest==8.1.1
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2024.4.16
rembg==2.0.56
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.3
scikit-image==0.23.1
scipy==1.13.0
semantic-version==2.10.0
sentry-sdk==1.45.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tensorstore==0.1.56
tifffile==2024.4.18
timm==0.9.16
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.1
tqdm==4.66.2
traitlets==5.14.3
transformers==4.25.1
transforms3d==0.4.1
trimesh==4.3.1
typing_extensions==4.11.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.29.0
vispy==0.14.2
wandb==0.16.6
wcwidth==0.2.13
websockets==11.0.3
zipp==3.18.1
